+++ app/static/lang/lang.en.js
@@ -451,6 +451,7 @@ export const lang = {
   loadWebAnnotationMessage2: { text: 'please try again' },
   noAnnotationsToDisplay: { text: 'No annotations to display' },
   flipPageToAnnotationText: { description: 'Flip page to this annotation' },
+  describeMarkup: { description: 'Describe this' },
   deleteAnnotation: { description: 'Delete this annotation' },
   deleteAnnotationConfirmation: { text: 'Are you sure you wish to delete this annotation?' },
   makeStandOffAnnotation: {